+++ e2e/runner/sui.go
@@ -117,6 +117,36 @@ func (r *E2ERunner) SuiWithdrawFungibleToken(
 	return tx
 }
 
+// SuiWithdrawAndCallFungibleToken calls WithdrawAndCall of Gateway with Sui fungible token ZRC20 on ZEVM
+func (r *E2ERunner) SuiWithdrawAndCallFungibleToken(
+	receiver string,
+	amount *big.Int,
+	payload sui.CallPayload,
+	revertOptions gatewayzevm.RevertOptions,
+) *ethtypes.Transaction {
+	receiverBytes, err := hex.DecodeString(receiver[2:])
+	require.NoError(r, err, "receiver: "+receiver[2:])
+
+	payloadBytes, err := payload.PackABI()
+	require.NoError(r, err)
+
+	tx, err := r.GatewayZEVM.WithdrawAndCall0(
+		r.ZEVMAuth,
+		receiverBytes,
+		amount,
+		r.SuiTokenZRC20Addr,
+		payloadBytes,
+		gatewayzevm.CallOptions{
+			IsArbitraryCall: false,
+			GasLimit:        big.NewInt(20000),
+		},
+		revertOptions,
+	)
+	require.NoError(r, err)
+
+	return tx
+}
+
 // SuiDepositSUI calls Deposit on Sui
 func (r *E2ERunner) SuiDepositSUI(
 	receiver ethcommon.Address,
@@ -206,6 +236,26 @@ func (r *E2ERunner) SuiMintUSDC(
 	require.NoError(r, err)
 
 	return r.suiExecuteTx(signer, tx)
+}
+
+// SuiCreateExampleWACPayload creates a payload for on_call function in Sui the example package
+// The example on_call function will just forward the withdrawn token to given 'suiAddress'
+func (r *E2ERunner) SuiCreateExampleWACPayload(suiAddress string) (sui.CallPayload, error) {
+	// only the CCTX's coinType is needed, no additional arguments
+	argumentTypes := []string{}
+	objects := []string{
+		r.SuiExample.GlobalConfigID.String(),
+		r.SuiExample.PartnerID.String(),
+		r.SuiExample.ClockID.String(),
+	}
+
+	// create the payload message from the sui address
+	message, err := hex.DecodeString(suiAddress[2:]) // remove 0x prefix
+	if err != nil {
+		return sui.CallPayload{}, err
+	}
+
+	return sui.NewCallPayload(argumentTypes, objects, message), nil
 }
 
 // SuiGetConnectedCalledCount reads the called_count from the GlobalConfig object in connected module